1.5.2 strategy interface 1

+++ src/OrleansRuntime/Scheduler/IWorkItem.cs
@@ -6,7 +6,7 @@ namespace Orleans.Runtime.Scheduler
     internal interface IWorkItem
     {
         // TODO: Change priority to general context
-        double PriorityContext { get; set; }
+        long PriorityContext { get; set; }
         ActivationAddress SourceActivation { get; set; }
         string Name { get; }
         WorkItemType ItemType { get; }
